Change next question nav to linear progression
Instead of computing the next unanswered question
we instead get the next question sorted list of questions
for the module, enabling the user to progress linear from
anywhere within the list of questions.

This is friendlier to users, allowing them to jump
forward and backward in the question list.

We preserve the calculation of the next unanswered
question method because that might be useful. We just grab
the next index in the question list from the current question
and check to make sure we are not exceeding the length of
the question list

Bump version to v0.9.0.dev38+devel