Add message, styling to authors why seeing imputed question
Users with authoring tool enbabled will see all
questions displayed as they navigate, even imputed questions.
This can be confusing to the user that an imputed question
appears. So we add a better message to remind the user
with Authoring Tool privilege enabled that they are seeing the
question for that reason.

Also bump VERSIOn and update CHANGELOG.