Remove legacy Organization from discussion invites, turn off discussion polling
Non-super users were not able to submit discussion comments
because the invitations being sent upon comment add still included
legacy organization reference. Invitations no longer have Organizations
in model after the removal of multi-tenant sub domains.

Removing the reference to organizations solved the issue with users
not being able to add comments.

We won't need organizations on discussion in future either, because
discussions will be tied to projects (or tasks) that still have
organization structure built-in.

Also turned of javascript initiation of polling for updates on
discussions in order to remove that noise from log. Polling was
a cheap way to check.