Fix help squad, reviewer settings management in 0.9.0. (#723)
Management of help squad and reviewers was tied to
Organization in 0.8.6. Javascript calls for adding/removing
users to help squad and reviewers assumed request passed organization
value.

For initial fix, we stop looking to request for organization
and set organization to the main organization.

Also add username to User.render_context_dict() method so
we can autocomplete on username.

Tweak Organization and User Admin list to see id and username respectively
in records list.