Require special env var to prevent accidental migrations to 0.9.0
Because Docker automatically runs database migrations, we add
the special environment variable DB_BACKED_UP_DO_UPGRADE that
must be intentionnally set to 'True' for dockerfile_exec.sh to
fully execute and migrate the database in deployments using Docker.

We test to make see if deployment's database is still in state
of version 0.8.6 and if so, then look for the DB_BACKED_UP_DO_UPGRADE
environment variable. If database has already been migrated, test
is ignored. If DB_BACKED_UP_DO_UPGRADE is set correctly to "True",
we proceed with migration.

This test avoids accidental migrations of databases without
organization first reading about upgrade to 0.9.0 and being
given the opportunity to back up database.