Add 'IT System' Django app for Wazuh integration
Create a new Django app "itsystems" to support new
models of "SystemInstance", "HostInstance", "AgentService"
and "Agent" in GovReady to facilitate exchanging data
with Wazuh and other scanning services that install agents
on hosts/servers/endpoints.

This is an attempt to model multi-server configurations
commonly associated with enterprise IT Systems.

SystemInstance - an individual IT System installed in a particular environment.
HostInstance - a "server" or other compute workload endpoint identifiable as part of a SystemInstance
AgentService - a monitoring/scanning service (e.g., Wazuh) that collects information from HostInstances
Agent - an instance of an agent installed on a HostInstance