Identify current_group in progress-history when id and module-id differ
This commit fixes an issue where the `id` and `module-id` of an app.yaml
question differ caused the progress history to fail display questions
properly in `task-progress-history-list.html` template.

Original matching query worked only when `id` and `module-id` were the
same, which is commonly the case by convention but not strictly true.

We need to find a better walk of relationships to get the `module-id` to
which the current question belongs.

Also minor changes to the `task-progress-history-list.html` to improve
display and code indents for better code folding.