0.9.0.dev Performance improvements (#774)
* Create migrations for shortening varchar field length

* remove obsolete-appearing siteapp/tests.orig.py

* need a sleep to get the node rendered

* fix title checks, and work a little on var_sleeps for not found elements

* needed to click on interstitial save to get to actual question

* click 'Got it' where the tests weren't expecting it

* add another 'Got it', and fix a regex

The regex match worked if the test was run in isolation, but the URL changed when the whole test suite was run, causing an ERROR.  We removed the variable part of the URL, but added a '^' to make sure we were grabbing the right part of the URL.

* Identify current_group in progress-history when id and module-id differ

This commit fixes an issue where the `id` and `module-id` of an app.yaml
question differ caused the progress history to fail display questions
properly in `task-progress-history-list.html` template.

Original matching query worked only when `id` and `module-id` were the
same, which is commonly the case by convention but not strictly true.

We need to find a better walk of relationships to get the `module-id` to
which the current question belongs.

Also minor changes to the `task-progress-history-list.html` to improve
display and code indents for better code folding.

* Improve performance: Set `is_finished` to false in progress history

To dramatically improve performance, set `is_finished` check to False
when showing question or showing module review page.

Current `is_finished` exhibited poor performance causing
caches to be rebuilt when answers changed on existing questions.
We were checking `is_finished` when building the `task-progress-history-list`.

We need a better `is_finished` test to answer state quickly.

Only lost on user interface is not seeing a check mark for an entire
module in the `task-progress-history-list`.

Also, apply the fix of identify current_group in `task-progress-history-list`
when id and module-id differ to the module review page view, too.

Changes to be committed:
modified:   guidedmodules/models.py
modified:   guidedmodules/views.py

* Update VERSION, CHANGELOG

* typo fixes

* Add a delay to see if that will help a test fail

https://circleci.com/gh/GovReady/govready-q/2066

```
FAIL: test_invitations (siteapp.tests.GeneralTests)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/circleci/project/siteapp/tests.py", line 506, in test_invitations
    self.assertRegex(self.browser.title, "Next Question: The Question") # user is on the task page
AssertionError: Regex didn't match: 'Next Question: The Question' not found in ''
```

* test_invitations(): adjust sleeps and add some INFO: print()s

The var_sleep() for the self.click_element('#transfer-editorship') fault actually needed to be on the _previous_ self.click_element("#save-button"), otherwise the test didn't get past the interstitial, and got out of phase.

The INFO: print()s help see which cycle the failure is happening in.

* Add back check for 'questions.get(mq.id)'

In test_invitations(), questions.get(mq.id) was None (I think when inviting me3).  This caused a Server Error 500, which caused the page title to be blank.