Synchronizing in 0.9.0.dev Performance improvements (#774)
* Identify current_group in progress-history when id and module-id differ

This commit fixes an issue where the `id` and `module-id` of an app.yaml
question differ caused the progress history to fail display questions
properly in `task-progress-history-list.html` template.

Original matching query worked only when `id` and `module-id` were the
same, which is commonly the case by convention but not strictly true.

We need to find a better walk of relationships to get the `module-id` to
which the current question belongs.

Also minor changes to the `task-progress-history-list.html` to improve
display and code indents for better code folding.

* Improve performance: Set `is_finished` to false in progress history

To dramatically improve performance, set `is_finished` check to False
when showing question or showing module review page.

Current `is_finished` exhibited poor performance causing
caches to be rebuilt when answers changed on existing questions.
We were checking `is_finished` when building the `task-progress-history-list`.

We need a better `is_finished` test to answer state quickly.

Only lost on user interface is not seeing a check mark for an entire
module in the `task-progress-history-list`.

Also, apply the fix of identify current_group in `task-progress-history-list`
when id and module-id differ to the module review page view, too.

* Split the `context.update` into smaller parts so it is possible to add in
timing code