Fix appsource field error handling for Django 2.2
This commit fixes two problems to gracefully handle appsource
field validation errors when updating an appsource and adding
incorrect data.

One error was appsource page trying to render the widget for
for available apps where no available apps existed because
an appsource connection could not be made.

The other error appeared to be failing to render error messages
properly through Django forms (for admin) because we needed to json.dumps
a variable that was an OrderedDict and was previously passed without
dumping to a string formatted as JSON.