Support vertical rendering of datagrid question type in output
Add `render` key can be added and set to `vertical` to force the tabular data
to render in a vertical format with the fields listed vertically on the side
instead of horizontally. This is good for rendering information about a single,
multi-field entry such as and address or Point of Content.