Use OSCAL and pure Python for control catalog
Replace SecControl library and parsing older XML version
of 800-53 catalog with pure python parsing of modern
OSCAL version of catalogs.

Create a new script `oscal.py` and a `Catalog` class
and create look ups of controls.

Because OSCAL is hierarchal and recursive, we also generate
a flattened, simplified version of our control catalog.