WIP profiling, better jinja2_expression_compile_cache cache, better next-page redirect
This interim performance profiling improvement incorporates pyinstrument library
in the guidedmodules/views.task_view to help see what is going in viewing a question.
This inclusion requires pip install pyinstrument.

A major change in this commit is to create a cache for Jinja `expression_compile` routine.
This new cache is `jinja2_expression_compile_cache` and seems to have a noticable improvement
on maintaining a regular amount of time to render questions.

We also noticed several page reloads that were the result of page redirects being handled
based on content of the next page inside the question save form. Calculating the next page
to display, which accounts for skipping computed pages, is now handled server side as a function
that can be called.

PROBLEM: Since the current new routing skips pages with imputed results, it becomes impossible
to navigate through completed questions. If you jump from the finished page to an imputed question
you come back to the finish page. If you jump from the finished page to an answered question in
order to change the answer, you can change the answer, but you do not proceed to the next already
answered question linearly. Instead, the application logic tries to send you to the next answerable
question, which could take you back to the finished page.

Also removed links to imputed pages from the module finished page. The links were not working
so it deid not make sense to display them as links.