Add simple CommonControl model
Add simple CommonControl and CommonControlProvider models.

We are creating CommonControl model as first step to being able
to manage control implementation statements in GovReady.

Initial use story is as a user/developer wanting to import
a spreadsheet of common controls to have a better interface for
reading the common controls against the control statement and
drafting the application layer controls for hybrid controls.

The user wants an easier way to do this than a spreadsheet because
spreadsheets are hard to read.

We also want a way to import CommonControls into GovReady database.

We create `CliControlImporter` as first cut tool for reading
an excel spreadsheet with xlrd, converting the rows to simple
control object and importing into the database. Code snippets
included. Class and snippets only work for one spreadsheet format,
but some abstraction exists to make customization easier for other
formats.