[WIP] Add Ajax calls to save implementation statements
This commit puts in the Django and Javascript to save a statement
to the server via Ajax.

Each implementation statement panel and form is given a unique id
when added to the page to make sure the correct statment information
is passed to the server.

When component name is update, the panel title updates.

Add hidden fields to send the control id and the system name as part
of the form.