Add Statement, Element models; Save implementation statements
This commits finishes work started in commit 4e77ba531069832d0a2507ebfd32a0e41f2bf8df

This commit adds Statement and Element models as the core feature
to track component-to-control statements. The more general terms of
"Statement" and "Element" are used instead of "Control"and "Component".

An "implementation statement" is a statement (and attestation).

A many to many relationship is supported for Statements and Elements.

This commit finishes in the Django and Javascript to save a statement
to the server via Ajax.

To Do:
- Delete statements
- Tests written