Make PDF and image generation optional and configurable.
PDF generator library wkhtmltopdf has security issues wherein individuals could add
HTML references such as links or file references inside the documents
they are creating which the PDF Generator blindly interpets. This leads
to SSRF (Server Side Request Forgery) in which data is retrieved from
server and added to PDF by the PDG Generator. We are also encountering
issues with the sub-dependency of `libxslt` on older versions of CentOS
before CentOS 8.x raising CVE vulnerability issues with scanners. For these
reasons, PDF Generation is being a configurable setting.

Supported options are currently 'None' and 'wkhtmltopdf' for PDF and
'None' and 'wkhtmltoimage' for image generation. This leaves
existing `wkhtmltopdf` code in place while creating default to 'None'.

If no PDF generator is set, the option to download PDF is removed from
task_finished page. As a backup, if there is an attempt to generate
a PDF from the Tasks model, the method gracefully fails when no
PDF generator is set and produces a text file instead with a message
stating PDF is not configured.

If `wkhtmltoimage` or valid image generator is not available, thumbnails
cannot be generated for uploaded files, either. Gracefully avoid trying
to create a thumbnail

Support new `environment.json` parameters of `GR_PDF_GENERATOR`
and `GR_IMG_GENERATOR`.