Tell bandit to accept use of pandoc without full path
Misplaced '# nosec' reference in trying to avoid Bandit scan.
Placing '# nosec' at end of multi-line statement.

Bandit wants a full path as part of any sub-process. We want
people to run pandoc where it is installed in their path.
The risk is low because we hardcode "process" and not pass it
as a variable that can be adjusted. We tell bandit to ignore
this line.