Make PDF generation and use of wkhtmltopdf optional (#842)
* Bump version to 0.9.1.5

* Make PDF and image generation optional and configurable.

PDF generator library wkhtmltopdf has security issues wherein individuals could add
HTML references such as links or file references inside the documents
they are creating which the PDF Generator blindly interpets. This leads
to SSRF (Server Side Request Forgery) in which data is retrieved from
server and added to PDF by the PDG Generator. We are also encountering
issues with the sub-dependency of `libxslt` on older versions of CentOS
before CentOS 8.x raising CVE vulnerability issues with scanners. For these
reasons, PDF Generation is being a configurable setting.

Supported options are currently 'None' and 'wkhtmltopdf' for PDF and
'None' and 'wkhtmltoimage' for image generation. This leaves
existing `wkhtmltopdf` code in place while creating default to 'None'.

If no PDF generator is set, the option to download PDF is removed from
task_finished page. As a backup, if there is an attempt to generate
a PDF from the Tasks model, the method gracefully fails when no
PDF generator is set and produces a text file instead with a message
stating PDF is not configured.

If `wkhtmltoimage` or valid image generator is not available, thumbnails
cannot be generated for uploaded files, either. Gracefully avoid trying
to create a thumbnail

Support new `environment.json` parameters of `GR_PDF_GENERATOR`
and `GR_IMG_GENERATOR`.

* remove spurious merge

* typo fixes

* typo fixes

* change important instance of 'generation' to 'generator'

* Fixes to optional PDF Generation settings

Fixes issue to make sure PDF generation turns on/off properly
with environmental settings. Fixes references to wkhtmltoimage.

 Please enter the commit message for your changes. Lines starting

* Fix reference to wkhtmltoimage

* Use 'off' instead of None fo wkhtmltopdf

* Use .circleci config with limited UI tests

* Upgrade markupsafe to 1.1.1 to address an error in package.

See https://github.com/pallets/markupsafe/issues/57#issuecomment-597105876

Version of markupsafe and setup tool had an accidental conflict.

* PK changes from 2020-05-03 to ge-wkhtmltopdf-optional (#843)

* call 'pandoc' instead of '/usr/bin/pandoc'

* add build-arg to Dockerfile, and check for 'wkhtmltopdf' in test-finished.html

* don't install wkhtmltopdf by default, because of security concerns

* don't install wkhtmltopdf by default, because of security concerns

* Tell bandit to accept use of pandoc without full path

Bandit wants a full path as part of any sub-process. We want
people to run pandoc where it is installed in their path.
The risk is low because we hardcode "process" and not pass it
as a variable that can be adjusted. We tell bandit to ignore
this line.

* Tell bandit to accept use of pandoc without full path

Misplaced '# nosec' reference in trying to avoid Bandit scan.
Placing '# nosec' at end of multi-line statement.

Bandit wants a full path as part of any sub-process. We want
people to run pandoc where it is installed in their path.
The risk is low because we hardcode "process" and not pass it
as a variable that can be adjusted. We tell bandit to ignore
this line.

* Re-adjust location of pandoc '# nosec'

* adjust comment

Co-authored-by: Greg Elin <greg.elin@govready.com>
Co-authored-by: Peter Kaminski <peter.kaminski@govready.com>
Co-authored-by: Peter Kaminski <kaminski@istori.com>