0.9.0 profiling 2020-04-16 merge master 2 (#871)
* remove obsolete-appearing siteapp/tests.orig.py

* need a sleep to get the node rendered

* fix title checks, and work a little on var_sleeps for not found elements

* needed to click on interstitial save to get to actual question

* click 'Got it' where the tests weren't expecting it

* add another 'Got it', and fix a regex

The regex match worked if the test was run in isolation, but the URL changed when the whole test suite was run, causing an ERROR.  We removed the variable part of the URL, but added a '^' to make sure we were grabbing the right part of the URL.

* Identify current_group in progress-history when id and module-id differ

This commit fixes an issue where the `id` and `module-id` of an app.yaml
question differ caused the progress history to fail display questions
properly in `task-progress-history-list.html` template.

Original matching query worked only when `id` and `module-id` were the
same, which is commonly the case by convention but not strictly true.

We need to find a better walk of relationships to get the `module-id` to
which the current question belongs.

Also minor changes to the `task-progress-history-list.html` to improve
display and code indents for better code folding.

* Improve performance: Set `is_finished` to false in progress history

To dramatically improve performance, set `is_finished` check to False
when showing question or showing module review page.

Current `is_finished` exhibited poor performance causing
caches to be rebuilt when answers changed on existing questions.
We were checking `is_finished` when building the `task-progress-history-list`.

We need a better `is_finished` test to answer state quickly.

Only lost on user interface is not seeing a check mark for an entire
module in the `task-progress-history-list`.

Also, apply the fix of identify current_group in `task-progress-history-list`
when id and module-id differ to the module review page view, too.

Changes to be committed:
modified:   guidedmodules/models.py
modified:   guidedmodules/views.py

* Update VERSION, CHANGELOG

* typo fixes

* Add a delay to see if that will help a test fail

https://circleci.com/gh/GovReady/govready-q/2066

```
FAIL: test_invitations (siteapp.tests.GeneralTests)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/circleci/project/siteapp/tests.py", line 506, in test_invitations
    self.assertRegex(self.browser.title, "Next Question: The Question") # user is on the task page
AssertionError: Regex didn't match: 'Next Question: The Question' not found in ''
```

* test_invitations(): adjust sleeps and add some INFO: print()s

The var_sleep() for the self.click_element('#transfer-editorship') fault actually needed to be on the _previous_ self.click_element("#save-button"), otherwise the test didn't get past the interstitial, and got out of phase.

The INFO: print()s help see which cycle the failure is happening in.

* Add back check for 'questions.get(mq.id)'

In test_invitations(), questions.get(mq.id) was None (I think when inviting me3).  This caused a Server Error 500, which caused the page title to be blank.

* Fix adding appsource guide

* WIP profiling, better jinja2_expression_compile_cache cache, better next-page redirect

This interim performance profiling improvement incorporates pyinstrument library
in the guidedmodules/views.task_view to help see what is going in viewing a question.
This inclusion requires pip install pyinstrument.

A major change in this commit is to create a cache for Jinja `expression_compile` routine.
This new cache is `jinja2_expression_compile_cache` and seems to have a noticable improvement
on maintaining a regular amount of time to render questions.

We also noticed several page reloads that were the result of page redirects being handled
based on content of the next page inside the question save form. Calculating the next page
to display, which accounts for skipping computed pages, is now handled server side as a function
that can be called.

PROBLEM: Since the current new routing skips pages with imputed results, it becomes impossible
to navigate through completed questions. If you jump from the finished page to an imputed question
you come back to the finish page. If you jump from the finished page to an answered question in
order to change the answer, you can change the answer, but you do not proceed to the next already
answered question linearly. Instead, the application logic tries to send you to the next answerable
question, which could take you back to the finished page.

Also removed links to imputed pages from the module finished page. The links were not working
so it deid not make sense to display them as links.

* WIP continue better next question calc to improve performance

Fixes a calculation problem in previous commit
a6e7a13f47f5964614ad903212bdc9f04250cc81 in which next question calculation
skipped questions with answered. This created an odd user experience of not
being able to answer/update a question and then proceed linearly. New calculation
includes all "answerable" questions while excluding the current question just
answered as an answerable question.

* Remove unwanted control files under development

* Precalc flattened controls to improve OSCAL control caching

Significantly improve page load time of project page
and question ages with improved caching of OSCAL control
catalogs.

Improvement comes from precalculating the flattened versions
of controls in a catalog and adding as a property to catalog
so values are part of cached singleton instance of catalog.

WARNING TODO: This precalculation along with instance caching of controls
may cause a problem in multi-tenant environment where different tenants have
have different organizational defined parameters.

* Bump VERSION

Co-authored-by: Peter Kaminski <kaminski@istori.com>
Co-authored-by: Greg Elin <greg.elin@govready.com>
Co-authored-by: Peter Kaminski <peter.kaminski@govready.com>