Make System object available to output templates
Add 'system' tag as another item that can be accessed
in the templates. The `{{ system }}` tag will access all
data properties available to the System data model, including
the controls statements for a system.

This is an initial implementation of the system tag and available
properties are likely to change.