Sync with master 0.9.1.22.4 - refesh output doc button
Add refresh output doc button to task finished pages (#928)

Add button to output documents on the task finish page to
allow users to force refresh of the output documents.

This will clear the cache of all the task.

User may have updated system information
and the result of the system update has been cached somewhere in the stack of
output documents that compose answers to a value in this document.
We cannot always tell what update has caused a cache of an output document to be dirty.