Closes #951. Closes #953. Assign non-admin project perms on system, root_element
When a new project is created by a non-admin, correctly assign the
permissions to the system and root_element to the user who started the project.

When a non-admin user invites a non-admin user to a join a project,
correctly assign edit perms to system and root_element.

Update documentation logging for new logged actions related to the above.

Fix mislabeled JSON response preventing full remove deleted statement
from editor.